Eliminated redundancy in storing an array of vertex pointers for each university
Created functions getPlayerVertexArray and destroyPlayerVertexArray which creates an array containing pointers to the vertices in the map in Game.c owned by the player, to replace the arrays stored permanently in the player. This means that data isn't redundant as this array is generated from map every time it is used.
Updated getExchangeRate to use the new vertexArray, to do this, I also made it properly structured with a single return at the end, since, in this course, functions may only return once at the end.
(This is important later in computing as it allows programs to be mathematically analysed since they are structured in a specific way)