Renamed size_t to size_t_local to fix errors when Anna compiled it
When Anna was trying to compile it, she got an error that size_t was
being redefined from stdlib, so I've renamed size_t to size_t_local.
Also commented out some of the repeated includes, hoping perhaps that
was what was preventing our program from compiling on open learning.